# src/api/swagger.yml
swagger: "2.0"
info:
  version: "0.0.1"
  title: TIER url shortener
  description: >-
    API to return shortened URLs
  contact:
    name: Max Tkachenko
    email: max.and.tkachenko@gmail.com
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths
basePath: /v1
schemes:
  - http
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /shorten:
    x-swagger-router-controller: shortUrls
    post:
      description: Shorten Url
      operationId: shorten
      parameters:
        - in: body
          name: url
          schema:
            type: object
            required:
              - longUrl
            properties:
              longUrl:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UrlShortenedResponse"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /{shortUrlHash}:
    x-swagger-router-controller: shortUrls
    get:
      description: Redirects to the long URL if there is one
      operationId: redirect
      produces:
        - text/html
      parameters:
        - name: shortUrlHash
          in: path
          description: short url hash
          required: true
          type: string
          minLength: 6
          maxLength: 6
      responses:
        "301":
          description: Success
          schema:
            type: string
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /stats/{shortUrlHash}:
    x-swagger-router-controller: shortUrls
    get:
      description: Returns visits statistics for specified shortened URL
      operationId: stats
      parameters:
        - name: shortUrlHash
          in: path
          description: short url hash
          required: true
          type: string
          minLength: 6
          maxLength: 6
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UrlShortenedStatsResponse"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
# complex objects have schema definitions
definitions:
# ----------------------------------
# success response schemas
# ----------------------------------
  UrlShortenedResponse:
    required:
      - status
      - data
    properties:
      status:
        $ref: "#/definitions/SuccessStatus"
      data:
        $ref: "#/definitions/UrlShortened"
  UrlShortenedStatsResponse:
    required:
      - status
      - data
    properties:
      status:
        $ref: "#/definitions/SuccessStatus"
      data:
        $ref: "#/definitions/UrlShortenedStats"
# ----------------------------------
# error response schemas
# ----------------------------------
  ErrorResponse:
    required:
      - status
      - message
    properties:
      status:
        $ref: "#/definitions/ErrorStatus"
      message:
        $ref: "#/definitions/ErrorMessage"
      code:
        $ref: "#/definitions/ErrorCode"
      data:
        description: >-
          A generic container for any other information about the
          error, i.e. the conditions that caused the error, stack
          traces, etc.
        type: object
# ----------------------------------
# complex object type schemas
# ----------------------------------
  UrlShortened:
    required:
      - shortUrl
      - longUrl
    properties:
      shortUrl:
        type: string
      longUrl:
        type: string

  UrlShortenedStats:
    required:
      - shortUrl
      - longUrl
      - visits
    properties:
      shortUrl:
        type: string
      longUrl:
        type: string
      visits:
        type: number

# ----------------------------------
# simple object type schemas
# ----------------------------------
  SuccessStatus:
    description: should always be set to "success"
    type: string
    enum:
      - success
  ErrorStatus:
    description: should always be set to "error"
    type: string
    enum:
      - error
  ErrorMessage:
    description: end-user-readable message, explaining what went wrong
    type: string
  ErrorCode:
    description: numeric code corresponding to the error
    type: number
    format: integer